requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-apscheduler==0.6.0
django-filter==21.1
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
